esp32-backend/requirements.txt
# ESP32-CAM Backend Server Dependencies
fastapi
uvicorn
opencv-python-headless
numpy
Pillow
requests
websockets
python-multipart
